Add files via upload
Just the fragment source files; you can merely add them to your project and connect the dots.  I'm using Google Maps on this one, so you'll need a Google API key for the map. Also, I had a HECK of a time early in the project, and found that I had to refactor my project as "AndroidX" to get it to work with Google Admob and Maps, due to excessive 'Manifest Merger' failures in a non-refactored project.